refactor: implement GetEmitters db func and update clean shutdown spam
